Move async task to own class
