+++ app/src/main/java/com/liggiorgio/rumapp/news/AsyncResponse.java
@@ -1,0 +1,7 @@
+package com.liggiorgio.rumapp.news;
+
+import java.util.ArrayList;
+
+public interface AsyncResponse {
+    void processFinish(ArrayList<String> output);
+}
